dividing Server module
