+++ dnsext-full-resolver/utils/benchmark.hs
@@ -24,7 +24,7 @@ import UnliftIO (concurrently_)
 
 import DNS.Cache.Iterative (Env (..))
 import qualified DNS.Cache.Iterative as Iterative
-import DNS.Cache.Server
+import DNS.Cache.Server.Bench
 import qualified DNS.Cache.TimeCache as TimeCache
 import qualified DNS.Do53.Memo as Cache
 import qualified DNS.Log as Log
